accessible/tests/mochitest/test_table_struct_ariagrid.html

<!DOCTYPE html>
<html>

<head>
  <title>Table accessible tree and table interface tests for ARIA grid</title>
  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/MochiKit/packed.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/common.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/role.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/table.js"></script>

  <script type="application/javascript">
    function doTest()
    {
      //////////////////////////////////////////////////////////////////////////
      // Pure ARIA grid
      var cellsArray = [
        [kColHeaderCell, kColHeaderCell, kColHeaderCell],
        [kRowHeaderCell, kDataCell,      kDataCell],
        [kRowHeaderCell, kDataCell,      kDataCell]
      ];

      testTableStruct("table", cellsArray);

      //////////////////////////////////////////////////////////////////////////
      // HTML table based ARIA grid
      var cellsArray = [
        [kColHeaderCell, kColHeaderCell, kColHeaderCell],
        [kDataCell,      kDataCell,      kDataCell],
        [kDataCell,      kDataCell,      kDataCell]
      ];

      testTableStruct("grid", cellsArray);

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addLoadEvent(doTest);
  </script>
</head>

<body>
  <a target="_blank"
     title="ARIA grid based on HTML table"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=491683">Mozilla Bug 491683</a>
  <a target="_blank"
     title="implement IAccessibleTable2"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=512424">Mozilla Bug 512424</a>

  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <!-- Not usual markup to avoid text accessible between cell accessibles -->
  <div id="table" role="grid">
    <div role="row"><span
      id="table_ch_1" role="columnheader">col_1</span><span
      id="table_ch_2" role="columnheader">col_2</span><span
      id="table_ch_3" role="columnheader">col_3</span></div>
    <div role="row"><span
      id="table_rh_1" role="rowheader">row_1</span><span
      id="table_dc_1" role="gridcell">cell1</span><span
      id="table_dc_2" role="gridcell">cell2</span></div>
    <div role="row"><span
      id="table_rh_2" role="rowheader">row_2</span><span
      id="table_dc_3" role="gridcell">cell3</span><span
      id="table_dc_4" role="gridcell">cell4</span></div>
  </div>

  <table role="grid" id="grid" border="1" cellpadding="10" cellspacing="0">
    <thead>
      <tr role="row">
        <th role="columnheader">subject</td>
        <th role="columnheader">sender</th>
        <th role="columnheader">date</th>
      </tr>
    </thead>
    <tbody>
      <tr role="row">
        <td role="gridcell" tabindex="0">about everything</td>
        <td role="gridcell">president</td>
        <td role="gridcell">today</td>
      </tr>
      <tr role="row">
        <td role="gridcell">new bugs</td>
        <td role="gridcell">mozilla team</td>
        <td role="gridcell">today</td>
      </tr>
    </tbody>
  </table>

</body>
</html>
